trajectory_generator/requirements.txt
numpy==1.18.1
matplotlib==3.3.1
trajectory_planning_helpers==0.76

# the following is required for minimum time optimization only
casadi==3.5.5
scipy==1.3.3
scikit-learn==0.23.1

# the following is required for path generator
opencv-contrib-python
pyyaml